INFUND-5630 Fixing failing acc tests.
Former-commit-id: 0b341e057321b42b5d37821c3c03d3175903ce1d

+++ ifs-data-service/src/main/java/com/worth/ifs/competition/transactional/CompetitionSetupQuestionServiceImpl.java
@@ -144,7 +144,7 @@ public class CompetitionSetupQuestionServiceImpl extends BaseTransactionalServic
                 }
                 formInputRepository.save(appendixFormInput);
             }
-        } else {
+        } else if (appendixFormInput != null) {
             formInputRepository.delete(appendixFormInput);
         }
     }
